Исправление ошибок

+++ ddl/initdb.sql
@@ -7,21 +7,21 @@ CREATE TABLE IF NOT EXISTS public.products(
 CREATE TABLE IF NOT EXISTS public.shop_mvideo(
     "date" date,
     product_id int,
-    sales_cnt money,
+    sales_cnt int,
     FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE SET NULL
 );
 
 CREATE TABLE IF NOT EXISTS public.shop_dns(
     "date" date,
     product_id int,
-    sales_cnt money,
+    sales_cnt int,
     FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE SET NULL
 );
 
 CREATE TABLE IF NOT EXISTS public.shop_sitilink(
     "date" date,
     product_id int,
-    sales_cnt money,
+    sales_cnt int,
     FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE SET NULL
 );
 
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS public.shop_name(
 CREATE TABLE IF NOT EXISTS public.plan(
     product_id int,
     id_shop int,
-    plan_cnt money,
+    plan_cnt int,
     plan_date date,
     FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE SET NULL,
     FOREIGN KEY (id_shop) REFERENCES shop_name (id_shop) ON DELETE SET NULL
